sysprof: replace `backtrace` with libunwind
`backtrace` fails to unwind the host stack in LuaJIT, since there are
no frame pointers during the VM calls. Sometimes, those failures cause
crashes. This commit replaces it with the libunwind-based unwinder,
which makes use of additional runtime information to provide robust
unwinding without any frame pointers. Because Tarantool uses its
own fiber unwinder, the new default unwinder is not built in
bundled builds.

Also, this commit enables C API tests, which used to crash with
`backtrace`, and disables an assertion regarding the successful
backtracer function setting, as it is unnecessary.

The `lj-603-err-snap-restore.test.lua` was updated to correspond with
the new size of the `arg` table.

Part of tarantool/tarantool#781