Implement use-def analysis for BC_VARG
Since use-def analysis for BC_VARG isn't implemented yet, some slots,
that should be considered as dead after BC_VARG recording, may be not
purged from the JIT slots. As a result, the assertion in the
`rec_check_slots()` fails.

This patch implement use-def analysis for BC_VARG considering that we
use:
* all slots from base (ra) to the maxslot, when we need to copy all
  vararg arguments (i.e. rb == 0)
* slots from base to base + n_results (rb + 1 == n_results >= 0) in
  other cases