Fix maxslot for recording of BC_TSETM
When recording TSETM instruction, maxslot are considering the same as
before it. But TSETM may release the top slot on Lua stack, that may
be used for the new frame during vmevent on recording. This leads to
the JIT slots inconsistency and assertion failure in
`record_check_slots()`.

This patch shrinks maxslot to the ra - 1 after TSETM recording, i.e.
the slot number of table, where the args are stored.