profilers: purge generation mechanism
Since both of the profiler parsers are now processing
the events in a stream-like fashion, the generation
mechanism is excessive and can be purged. This results
in a significant memory consumption drop, especially
for the AVL-tree part.

Consider this script:
| jit.off()
| misc.sysprof.start{mode = 'C', interval=10}
| for i = 1, 1e7 do tostring(i) end
| misc.sysprof.stop()

After executing it with LuaJIT, you can parse it like this:
| $ time -v luajit-parse-sysprof sysprof.bin

So, before the patch:
| Maximum resident set size (kbytes): 224928

And after the patch:
| Maximum resident set size (kbytes): 32780

That is the 85% reduction in memory consumption.

Follows up tarantool/tarantool#8700